requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.0.3
djangorestframework 3.15.1
pip                 22.0.4
setuptools          58.1.0
sqlparse            0.4.4
typing_extensions   4.10.0
